Harden DI and serializer edge cases
