+++ docs/plans/2026-07-12-modern-di-stj-plan.md
@@ -40,19 +40,20 @@
 ## Execution results
 
 - Baseline: 300 core tests and 10 MessagePack tests passed; 18 existing tests were skipped.
-- Final non-Windows suite: 370 core tests and 12 MessagePack tests passed; 0 failed; 18 existing tests were skipped.
-- Focused DI coverage now includes singleton and transient compatibility, constructor injection, isolated containers, coherent provider snapshots after late replacement, `IServiceCollection` overrides and enumeration, open generics, exact disposal of every provider snapshot, external instance ownership, per-host client isolation, and host-versus-caller disposal ownership.
-- Serializer coverage retains exact wire-format assertions and adds regressions for exclusions inside collections and settings, named floating-point values, declared-type depth limits, cycles, malformed getters, predicates, converters, and iterators, STJ custom converters/number handling/extension data, stream/string parity, public fields, the legacy ignore attribute, every `DataDictionary` interface path, structured/raw values, settings and POST-data coercion, and MessagePack raw-object round-trips.
-- Linux CI now collects and uploads Cobertura coverage for both test assemblies. The final core report is 74% line / 50% branch overall and 88.1% of executable production lines changed by the PR are covered; the migration-critical components are substantially higher: `JsonValueWriter` 97% line / 93% branch, `DataDictionaryConverter` 93% / 81%, `JsonElementValueConverter` 96% / 96%, `SettingsDictionaryConverter` 89% / 83%, `DataDictionary` 99% / 92%, and the Microsoft DI resolver 84% / 61%.
+- Final non-Windows suite: 377 core tests and 13 MessagePack tests passed; 0 failed; 18 existing tests were skipped.
+- Focused DI coverage now includes singleton and transient compatibility, constructor injection, isolated containers, coherent provider snapshots after late replacement, `IServiceCollection` overrides and enumeration, open generics, deterministic constructor/factory cycle detection without false positives across concurrent provider resolutions, captured `IServiceProvider` behavior, exact disposal of every provider snapshot, external instance ownership, per-host client isolation, and host-versus-caller disposal ownership.
+- Serializer coverage retains exact wire-format assertions and adds regressions for exclusions inside collections and settings, named floating-point values, declared-type depth limits, cycles, malformed getters, predicates, converters, and iterators, initialized source-generated contexts, STJ custom converters/number handling/extension data, stream/string parity, public fields, the legacy ignore attribute, every `DataDictionary` interface path, structured/raw values, failed-object diagnostics, null/unstructured POST data, settings coercion, and MessagePack storage round-trips.
+- Linux CI now collects and uploads Cobertura coverage for both test assemblies. The final core report is 74.2% line / 50.1% branch overall and 88.9% of executable production lines changed by the PR are covered; the migration-critical components are substantially higher: `JsonValueWriter` 97.2% line / 92.5% branch, `DataDictionaryConverter` 92.6% / 81.3%, `JsonElementValueConverter` 95.7% / 95.7%, `PostDataConverter` 95.8% / 72.2%, `SettingsDictionaryConverter` 88.9% / 83.3%, `DataDictionary` 98.9% / 92.0%, and the Microsoft DI resolver/fallback 85.8% / 64.5% and 89.5% / 80.0% respectively.
 - `net462` core and `net472` test assemblies build with 0 warnings and 0 errors when Windows targeting is forced on macOS. Runtime execution of the .NET Framework tests still belongs in Windows CI.
-- The non-Windows solution packs successfully. A Windows-shaped core package containing both `net462` and `netstandard2.0` assets also packs successfully with explicit `SolutionDir`.
-- `net8.0` and `net10.0` core builds pass with 0 AOT/trim/single-file warnings. Real NativeAOT executables now cover the packed NuGet assets, default Microsoft DI/default serialization, source-generated custom properties and fields, event-batch serialization, storage round-trips, queued submission, identifiable exception frames across rethrow boundaries, and Generic Host startup/shutdown on both target frameworks. The `net10.0` smoke found and fixed an offset-only runtime frame with no usable identity; those empty frames are now omitted. Linux CI publishes and executes the package and hosting smoke applications with linker warnings treated as errors.
+- The non-Windows solution packs successfully, with the root Markdown documentation declared as the modern NuGet package readme while retaining the legacy `readme.txt`. A Windows-shaped core package containing both `net462` and `netstandard2.0` assets also packs successfully with explicit `SolutionDir`.
+- `net8.0` and `net10.0` core builds pass with 0 AOT/trim/single-file warnings. Real NativeAOT executables now cover the packed NuGet assets, default Microsoft DI/default serialization, built-in configuration deserialization, source-generated custom properties and fields, compatibility-ignore filtering, event-batch serialization, storage round-trips, queued submission, identifiable exception frames across rethrow boundaries, and Generic Host startup/shutdown on both target frameworks. The `net10.0` smoke found and fixed an offset-only runtime frame with no usable identity; those empty frames are now omitted. Linux CI publishes and executes the package and hosting smoke applications with ILLink and trim-analysis warnings treated as errors.
 - SDK ApiCompat is clean from the previous `netstandard2.0` asset to both the new `netstandard2.0` and `net8.0` assets. Compatibility shims retain the legacy ignore attribute, serializer delegates, and `AssemblyHelper.GetTypes` API.
+- The current direct and transitive NuGet vulnerability audit is clean across every production and test project.
 - Release builds now actually treat warnings as errors; the previous `WarningsAsErrors=true` property was a warning-code list rather than the boolean build gate.
 
 ## Release impact
 
-- The non-Windows core package grows from 596,508 bytes to 686,321 bytes (about 15%) because it contains `netstandard2.0`, `net8.0`, and `net10.0` assets instead of one asset. Its uncompressed contents shrink from 2,261,763 bytes to 1,792,690 bytes (about 21%), and the `netstandard2.0` assembly shrinks from 998,912 bytes to 348,160 bytes (about 65%).
+- The non-Windows core package grows from 596,508 bytes to 699,874 bytes (about 17%) because it contains `netstandard2.0`, `net8.0`, and `net10.0` assets plus the modern package readme instead of one asset. Its uncompressed contents shrink from 2,261,763 bytes to 1,824,641 bytes (about 19%), and the `netstandard2.0` assembly shrinks from 998,912 bytes to 354,304 bytes (about 65%).
 - Every package inherits `Microsoft.Extensions.DependencyInjection` 8 through the core package. The `netstandard2.0`, `net462`, and `net8.0` dependency groups also carry System.Text.Json 10, which is the largest deployment and binding-risk change for .NET Framework applications.
 - The existing `IDependencyResolver` and public setup APIs remain compatible. Microsoft DI backs the client-owned resolver and exposes an `IServiceCollection` customization seam; registrations from an application's root Generic Host provider are not automatically imported into the client's isolated provider.
 - Configuration-based hosting overloads now create one client owned by each host provider instead of mutating the process-wide `ExceptionlessClient.Default`. Explicit client overloads remain caller-owned. Configuration-based logging providers likewise own an isolated client, while providers given an existing client only flush it and do not dispose it. This removes cross-host configuration/disposal interference but changes code that intentionally depended on those overloads mutating the global default.
