+++ test/Exceptionless.Tests/Dependency/DependencyTests.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
 using Exceptionless.Dependency;
 using Exceptionless.Serializer;
 using Microsoft.Extensions.DependencyInjection;
@@ -142,6 +144,35 @@ namespace Exceptionless.Tests.Dependency {
         }
 
         [Fact]
+        public void Resolve_WithCircularDependencies_ThrowsInvalidOperationException() {
+            // Arrange
+            using var resolver = new DefaultDependencyResolver();
+            resolver.Register<ICircularServiceA, CircularServiceA>();
+            resolver.Register<ICircularServiceB, CircularServiceB>();
+
+            // Act
+            Exception exception = Record.Exception(() => resolver.Resolve<ICircularServiceA>());
+
+            // Assert
+            var invalidOperationException = Assert.IsType<InvalidOperationException>(exception);
+            Assert.Contains("circular dependency", invalidOperationException.Message, StringComparison.OrdinalIgnoreCase);
+        }
+
+        [Fact]
+        public void Resolve_WithCircularFactory_ThrowsInvalidOperationException() {
+            // Arrange
+            using var resolver = new DefaultDependencyResolver();
+            resolver.Register(typeof(IServiceA), () => resolver.Resolve<IServiceA>());
+
+            // Act
+            Exception exception = Record.Exception(() => resolver.Resolve<IServiceA>());
+
+            // Assert
+            var invalidOperationException = Assert.IsType<InvalidOperationException>(exception);
+            Assert.Contains("circular dependency", invalidOperationException.Message, StringComparison.OrdinalIgnoreCase);
+        }
+
+        [Fact]
         public void Resolve_WithConcreteRegistration_ReturnsTransientInstances() {
             // Arrange
             var resolver = new DefaultDependencyResolver();
@@ -156,6 +187,44 @@ namespace Exceptionless.Tests.Dependency {
         }
 
         [Fact]
+        public async Task Resolve_WithConcurrentProviderResolution_DoesNotReportCircularDependency() {
+            // Arrange
+            var services = new ServiceCollection();
+            services.AddSingleton<ServiceProviderConsumer>();
+            using var resolver = new DefaultDependencyResolver(services);
+            using var firstEntered = new ManualResetEventSlim();
+            using var bothEntered = new CountdownEvent(2);
+            using var release = new ManualResetEventSlim();
+            CancellationToken cancellationToken = TestContext.Current.CancellationToken;
+            int activations = 0;
+            resolver.Register(typeof(IServiceA), () => {
+                if (Interlocked.Increment(ref activations) == 1)
+                    firstEntered.Set();
+
+                bothEntered.Signal();
+                release.Wait(TimeSpan.FromSeconds(5), cancellationToken);
+                return new ServiceA();
+            });
+            IServiceProvider serviceProvider = resolver.Resolve<ServiceProviderConsumer>().ServiceProvider;
+
+            // Act
+            Task<object> first = Task.Run(() => serviceProvider.GetService(typeof(IServiceA)));
+            bool firstEnteredInTime = firstEntered.Wait(TimeSpan.FromSeconds(5), cancellationToken);
+            Task<object> second = Task.Run(() => serviceProvider.GetService(typeof(IServiceA)));
+            bool enteredConcurrently = bothEntered.Wait(TimeSpan.FromSeconds(5), cancellationToken);
+            release.Set();
+            object[] resolved = null;
+            Exception exception = await Record.ExceptionAsync(async () => resolved = await Task.WhenAll(first, second));
+
+            // Assert
+            Assert.True(firstEnteredInTime);
+            Assert.True(enteredConcurrently);
+            Assert.Null(exception);
+            Assert.Equal(2, activations);
+            Assert.All(resolved, Assert.NotNull);
+        }
+
+        [Fact]
         public void Resolve_WithFactoryRegistration_ReturnsTransientInstances() {
             // Arrange
             var resolver = new DefaultDependencyResolver();
@@ -167,6 +236,20 @@ namespace Exceptionless.Tests.Dependency {
 
             // Assert
             Assert.NotSame(first, second);
+        }
+
+        [Fact]
+        public void Resolve_WithIServiceProviderDependency_ReturnsUsableProvider() {
+            // Arrange
+            using var resolver = new DefaultDependencyResolver();
+            resolver.Register<IServiceA, ServiceA>();
+
+            // Act
+            var consumer = resolver.Resolve<ServiceProviderConsumer>();
+            var service = consumer.ServiceProvider.GetService(typeof(IServiceA));
+
+            // Assert
+            Assert.Same(resolver.Resolve<IServiceA>(), service);
         }
 
         [Fact]
@@ -249,6 +332,14 @@ namespace Exceptionless.Tests.Dependency {
 
     public class ServiceC {}
 
+    public class ServiceProviderConsumer {
+        public ServiceProviderConsumer(IServiceProvider serviceProvider) {
+            ServiceProvider = serviceProvider;
+        }
+
+        public IServiceProvider ServiceProvider { get; }
+    }
+
     public interface IGenericService<T> { }
 
     public class GenericService<T> : IGenericService<T> { }
@@ -275,5 +366,17 @@ namespace Exceptionless.Tests.Dependency {
         public void Dispose() {
             DisposeCount++;
         }
+    }
+
+    public interface ICircularServiceA { }
+
+    public interface ICircularServiceB { }
+
+    public class CircularServiceA : ICircularServiceA {
+        public CircularServiceA(ICircularServiceB service) { }
+    }
+
+    public class CircularServiceB : ICircularServiceB {
+        public CircularServiceB(ICircularServiceA service) { }
     }
 }
